[FIX]Fixed the review points.
