+++ web_external_help/models/__init__.py
@@ -1,4 +1,5 @@
 # See LICENSE file for full copyright and licensing details.
 
 from . import fields_external_help
+from . import field_help_line
 from . import model_external_help
