+++ web_external_help/security/res_groups.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="UTF-8"?>
 <odoo>
     <record id="group_help_manager" model="res.groups">
         <field name="name">Help Manager</field>
