+++ web_external_help/views/debug_menu.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="UTF-8"?>
 
 <templates id="template" xml:space="preserve">
     <t t-extend="WebClient.DebugManager.Global">
