+++ web_external_help/views/view_model_external_help.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="UTF-8"?>
 <odoo>
     <record id="model_external_help_view_form" model="ir.ui.view">
         <field name="name">model.external.help.form</field>
